// broccoliapps.com/src/db/schemas.ts
import { table } from "@broccoliapps/backend";

export type AuthCode = {
  code: string;
  app: string;
  name: string;
  email: string;
  userId: string;
  provider: string;
  expiresAt: number;
  ttl: number;
};

export const authCodes = table<AuthCode>("authCode").key(["code"]).build();

export type MagicLinkToken = {
  token: string;
  email: string;
  app: string;
  platform?: string;
  createdAt: number;
  expiresAt: number;
  ttl: number;
};

export const magicLinkTokens = table<MagicLinkToken>("magicLinkToken").key(["token"]).build();